Refactor: CompanyDetailsCard removed, and logic transferred to DetailsPage.
