+++ pages/companies/DetailsPage.module.css
@@ -1,0 +1,110 @@
+.infos {
+  padding: 5px;
+  background-color: white;
+}
+
+.companyName {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  padding: 15px 0;
+}
+
+.chevron {
+  margin-left: -7px;
+  margin-bottom: -3px;
+}
+
+.iframeSection {
+  width: 100%;
+  min-height: 15rem;
+  margin-bottom: 21px;
+  border-style: none;
+}
+
+.closeIframeButton {
+  display: flex;
+  justify-content: center;
+}
+
+.buttonsBar {
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+}
+
+@media screen and (min-width: 375px) {
+  .companyName {
+    padding: 21px 0;
+  }
+
+  .infos {
+    padding: 10px;
+  }
+
+  .iframeSection {
+    min-height: 22rem;
+  }
+
+  .buttonsBar {
+    flex-direction: row;
+    justify-content: space-around;
+  }
+}
+
+@media screen and (min-width: 500px) {
+  .companyName {
+    margin-left: 2rem;
+    padding: 42px 0;
+    justify-content: left;
+  }
+  .infos {
+    border-radius: 8px;
+    padding: 21px;
+    margin: 0 2rem;
+  }
+
+  .iframeSection {
+    min-height: 28rem;
+  }
+
+  .closeIframeButton {
+    flex-direction: column;
+    align-items: baseline;
+  }
+
+  .buttonsBar {
+    flex-direction: column;
+    align-items: baseline;
+  }
+}
+
+@media screen and (min-width: 764px) {
+  .companyName {
+    margin-left: 4rem;
+    padding: 50px 0;
+  }
+  .infos {
+    margin: 0 4rem 6rem;
+    padding: 42px;
+  }
+
+  .iframeSection {
+    min-height: 33rem;
+  }
+
+  .buttonsBar {
+    gap: 1rem;
+  }
+}
+
+@media screen and (min-width: 1220px) {
+  .companyName {
+    margin-left: 6rem;
+  }
+  .infos {
+    margin: 0 6rem 8rem;
+    padding: 60px;
+  }
+}
